Introducing TundraCore: Asset, Audio, Console, Framework, Input, Scene, and Ui static libraries now combined into one dynamic library, TundraCore. Alternatively, TundraCore can be built as a single static library by setting TUDRACORE_SHARED to 0 in the root CMakeLists.txt. As the result, depending on the case, the resulting plugin DLL's size can be even three times smaller compared to the non-TundraCore approach. Comparison done using RelWithDebInfo build configuration.
The TundraCore source folder has the original libraries' sources in their own subdirectories. Use UseTundraCore() CMake macro
to to include all of the TundraCore subdirectories to your project in order to include the files using the shorter include path
(f.ex. "Ui/UiAPI.h" vs. "UiApi.h"). Probably ideally this macro should not be needed at all, and all of the folder structure
should be included automatically.

Note: For now the Color class resides in Math library due to problematic nature of compile-time MATH_OGRE_INTEROP feature
which creates dependency to Ogre and we don't want TundraCore to have Ogre dependency. Also note that, for now at least, Math, kNet and Qt needs to be linked explicitly to the project in most cases.

Also contains build fixes in SceneWidgetComponents and VlcPlugin related to the SceneInteract refactoring.